+++ aiobungie/crate/fireteams.py
@@ -90,6 +90,9 @@ class FireteamActivity(int, enums.Enum):
     S14_EXPUNGE = 30
     S15_ASTRAL_ALIGNMENT = 31
     S15_SHATTERED_RELAM = 32
+    SHATTERED_THRONE = 33
+    PROPHECY = 34
+    PIT_OF_HERESY = 35
     DOE = 36
     """Dares of Eternity."""
     DUNGEON_GOA = 37